[AN] refactor: 중복 되는 파일 및 코드 제거 #125
