+++ android/app/src/main/java/com/onair/hearit/presentation/library/LibraryFragment.kt
@@ -18,7 +18,6 @@ class LibraryFragment :
     @Suppress("ktlint:standard:backing-property-naming")
     private var _binding: FragmentLibraryBinding? = null
     private val binding get() = _binding!!
-    private val viewModel: BookmarkViewModel by viewModels { BookmarkViewModelFactory() }
 
     private val viewModel: LibraryViewModel by viewModels { LibraryViewModelFactory() }
 
@@ -56,11 +55,6 @@ class LibraryFragment :
     private fun observeViewModel() {
         viewModel.bookmarks.observe(viewLifecycleOwner) { bookmarks ->
             adapter.submitList(bookmarks)
-            observeViewModel()
-
-            // 테스트용으로 더미 데이터 넣어 놓음
-            val bookmarks = BookmarkDummyData.getBookmarks()
-            adapter.submitList(bookmarks)
 
             binding.layoutLibraryWhenNoBookmark.visibility =
                 if (bookmarks.isNullOrEmpty()) View.VISIBLE else View.GONE
@@ -69,9 +63,7 @@ class LibraryFragment :
         viewModel.toastMessage.observe(viewLifecycleOwner) { resId ->
             Toast.makeText(requireContext(), getString(resId), Toast.LENGTH_SHORT).show()
         }
-    }
 
-    private fun observeViewModel() {
         viewModel.uiState.observe(viewLifecycleOwner) { uiState ->
             binding.uiState = uiState
         }
