+++ android/app/src/main/java/com/onair/hearit/presentation/library/LibraryViewModelFactory.kt
@@ -3,7 +3,9 @@ package com.onair.hearit.presentation.library
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.onair.hearit.data.datasource.BookmarkRemoteDataSourceImpl
+import com.onair.hearit.data.datasource.MemberRemoteDataSourceImpl
 import com.onair.hearit.data.repository.BookmarkRepositoryImpl
+import com.onair.hearit.data.repository.MemberRepositoryImpl
 import com.onair.hearit.di.NetworkProvider
 
 @Suppress("UNCHECKED_CAST")
@@ -13,6 +15,9 @@ class LibraryViewModelFactory : ViewModelProvider.Factory {
             BookmarkRemoteDataSourceImpl(NetworkProvider.bookmarkService)
         val bookmarkRepository = BookmarkRepositoryImpl(bookmarkRemoteDataSource)
 
-        return LibraryViewModel(bookmarkRepository) as T
+        val memberRemoteDataSource = MemberRemoteDataSourceImpl(NetworkProvider.memberService)
+        val memberRepository = MemberRepositoryImpl(memberRemoteDataSource)
+
+        return LibraryViewModel(bookmarkRepository, memberRepository) as T
     }
 }
